Moved navigation package to root
