+++ app/src/main/java/tech/gusgol/portabletrainer/ui/workouts/create/CreateWorkoutRoute.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
-import tech.gusgol.portabletrainer.ui.navigation.PortableTrainerDestinations
+import tech.gusgol.portabletrainer.navigation.PortableTrainerDestinations
 import tech.gusgol.portabletrainer.ui.workouts.navigation.WorkDetailDestination
 
 @Composable
